Feat: new features
It was changed the url for http request when app request all files and request a file by id in the file files.services.js. It were some changes for integrate these new features in the files; DocumentManager.jsx and VisualizerDocument.jsx. The last thing was delete a const in the file static.data.js