Feat: new dependency , feature and component
It was installed moment dependency and implemented in the component TableFiles.jsx . It was create a new css style called TabledFiles.css which was implemented in TableFiles.jsx also in this component were implemented new features. In the component DocumentManager.jsx were implemente new css style. The last thing was added a new variable in the index.css file