Feat:  new file and features
It was create an .env file in the source of the project to save an environment variable, which was implemented in the file axios.js. In axios.js were implemented new axios features for each occasions when user make a request. In files.services.js were changed the endpoints for new axios functions